updating testing branch

+++ src/abs/model/Customer.java
@@ -4,19 +4,19 @@
 package abs.model;
 
 /**
- * @author Daniel Caddaye
+ * 
  *
  */
 public class Customer extends AbstractUser {
 
 	private String address;
-	private int phone;
+	private double phone;
 
 	/**
 	 * 
 	 */
 
-	public Customer(String name, String email, String address, int phone, String password) {
+	public Customer(String name, String email, String address, double phone, String password) {
 		super(name, email, address);
 		this.address = address;
 		this.phone = phone;
